+++ sdk/communication/azure-communication-callautomation/azure/communication/callautomation/_generated/aio/_client.py
@@ -18,16 +18,11 @@ from azure.core.rest import AsyncHttpResponse, HttpRequest
 from .. import models as _models
 from .._utils.serialization import Deserializer, Serializer
 from ._configuration import AzureCommunicationCallAutomationServiceConfiguration
-from .operations import (
-    AzureCommunicationCallAutomationServiceOperationsMixin,
-    CallConnectionOperations,
-    CallDialogOperations,
-    CallMediaOperations,
-    CallRecordingOperations,
-)
+from .operations import CallConnectionOperations, CallMediaOperations, CallRecordingOperations
+from .operations._operations import _AzureCommunicationCallAutomationServiceOperationsMixin
 
 
-class AzureCommunicationCallAutomationService(AzureCommunicationCallAutomationServiceOperationsMixin):
+class AzureCommunicationCallAutomationService(_AzureCommunicationCallAutomationServiceOperationsMixin):
     """Azure Communication Service Call Automation APIs.
 
     :ivar call_connection: CallConnectionOperations operations
@@ -35,8 +30,6 @@ class AzureCommunicationCallAutomationService(AzureCommunicationCallAutomationSe
      azure.communication.callautomation.aio.operations.CallConnectionOperations
     :ivar call_media: CallMediaOperations operations
     :vartype call_media: azure.communication.callautomation.aio.operations.CallMediaOperations
-    :ivar call_dialog: CallDialogOperations operations
-    :vartype call_dialog: azure.communication.callautomation.aio.operations.CallDialogOperations
     :ivar call_recording: CallRecordingOperations operations
     :vartype call_recording:
      azure.communication.callautomation.aio.operations.CallRecordingOperations
@@ -44,8 +37,8 @@ class AzureCommunicationCallAutomationService(AzureCommunicationCallAutomationSe
     :type endpoint: str
     :param credential: Credential needed for the client to connect to Azure. Required.
     :type credential: ~azure.core.credentials.AzureKeyCredential
-    :keyword api_version: Api Version. Default value is "2024-09-01-preview". Note that overriding
-     this default value may result in unsupported behavior.
+    :keyword api_version: Api Version. Default value is "2025-05-15". Note that overriding this
+     default value may result in unsupported behavior.
     :paramtype api_version: str
     """
 
@@ -81,7 +74,6 @@ class AzureCommunicationCallAutomationService(AzureCommunicationCallAutomationSe
         self._serialize.client_side_validation = False
         self.call_connection = CallConnectionOperations(self._client, self._config, self._serialize, self._deserialize)
         self.call_media = CallMediaOperations(self._client, self._config, self._serialize, self._deserialize)
-        self.call_dialog = CallDialogOperations(self._client, self._config, self._serialize, self._deserialize)
         self.call_recording = CallRecordingOperations(self._client, self._config, self._serialize, self._deserialize)
 
     def send_request(