+++ sdk/purestorageblock/azure-mgmt-purestorageblock/generated_samples/avs_storage_containers_list_by_storage_pool_maximum_set_gen.py
@@ -38,6 +38,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: 2024-11-01-preview/AvsStorageContainers_ListByStoragePool_MaximumSet_Gen.json
+# x-ms-original-file: 2024-11-01/AvsStorageContainers_ListByStoragePool_MaximumSet_Gen.json
 if __name__ == "__main__":
     main()