+++ sdk/purestorageblock/azure-mgmt-purestorageblock/generated_samples/storage_pools_list_by_subscription_maximum_set_gen.py
@@ -35,6 +35,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: 2024-11-01-preview/StoragePools_ListBySubscription_MaximumSet_Gen.json
+# x-ms-original-file: 2024-11-01/StoragePools_ListBySubscription_MaximumSet_Gen.json
 if __name__ == "__main__":
     main()