+++ sdk/resources/azure-mgmt-resource/generated_samples/policy/create_policy_assignment_non_compliance_messages.py
@@ -57,6 +57,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/resources/resource-manager/Microsoft.Authorization/stable/2022-06-01/examples/createPolicyAssignmentNonComplianceMessages.json
+# x-ms-original-file: specification/resources/resource-manager/Microsoft.Authorization/stable/2023-04-01/examples/createPolicyAssignmentNonComplianceMessages.json
 if __name__ == "__main__":
     main()