+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -12,7 +12,8 @@
   ([#41950](https://github.com/Azure/azure-sdk-for-python/pull/41950))
 - Customer Facing Statsbeat: Added logic for retry item count
   ([#41971](https://github.com/Azure/azure-sdk-for-python/pull/41971))
-- Customer Facing Statsbeat: Added logic to include the metrices in exporter and base classes
+- Customer Facing Statsbeat: Updated base and storage to include the metrics for customer statsbeat
+  ([#42123](https://github.com/Azure/azure-sdk-for-python/pull/42123))
 
 
 ## 1.0.0b39 (2025-06-25)